+++ plugins/Diagnostics/DiagnosticReport.php
@@ -7,21 +7,7 @@
  */
 namespace Piwik\Plugins\Diagnostics;
 
-use Piwik\ArchiveProcessor\Rules;
-use Piwik\CliMulti\CliPhp;
-use Piwik\Common;
-use Piwik\Config;
-use Piwik\CronArchive;
-use Piwik\Date;
-use Piwik\Db;
-use Piwik\DbHelper;
-use Piwik\Development;
-use Piwik\Option;
-use Piwik\Piwik;
-use Piwik\Plugin\Manager;
 use Piwik\Plugins\Diagnostics\Diagnostic\DiagnosticResult;
-use Piwik\SettingsPiwik;
-use Piwik\Version;
 
 /**
  * A diagnostic report contains all the results of all the diagnostics.
@@ -39,6 +25,11 @@ class DiagnosticReport
     private $optionalDiagnosticResults;
 
     /**
+     * @var DiagnosticResult[]
+     */
+    private $informationalResults;
+
+    /**
      * @var int
      */
     private $errorCount = 0;
@@ -51,11 +42,13 @@ class DiagnosticReport
     /**
      * @param DiagnosticResult[] $mandatoryDiagnosticResults
      * @param DiagnosticResult[] $optionalDiagnosticResults
+     * @param DiagnosticResult[] $informationalResults
      */
-    public function __construct(array $mandatoryDiagnosticResults, array $optionalDiagnosticResults)
+    public function __construct(array $mandatoryDiagnosticResults, array $optionalDiagnosticResults, array $informationalResults)
     {
         $this->mandatoryDiagnosticResults = $mandatoryDiagnosticResults;
         $this->optionalDiagnosticResults = $optionalDiagnosticResults;
+        $this->informationalResults = $informationalResults;
 
         $this->computeErrorAndWarningCount();
     }
@@ -97,7 +90,7 @@ class DiagnosticReport
      */
     public function getAllResults()
     {
-        return array_merge($this->mandatoryDiagnosticResults, $this->optionalDiagnosticResults);
+        return array_merge($this->mandatoryDiagnosticResults, $this->optionalDiagnosticResults, $this->informationalResults);
     }
 
     /**
@@ -118,146 +111,7 @@ class DiagnosticReport
 
     public function getInformationalResults()
     {
-        $results = [];
-
-        $results[] = DiagnosticResult::informationalResult('Matomo Version', Version::VERSION);
-
-        if ( ! empty( $_SERVER['SERVER_SOFTWARE'] ) ) {
-            $results[] = DiagnosticResult::informationalResult('Server Info', $_SERVER['SERVER_SOFTWARE']);
-        }
-        if ( defined('PHP_OS') && PHP_OS ) {
-            $results[] = DiagnosticResult::informationalResult('PHP_OS',  PHP_OS);
-        }
-        if ( defined('PHP_BINARY') && PHP_BINARY ) {
-            $results[] = DiagnosticResult::informationalResult('PHP_BINARY', PHP_BINARY);
-        }
-        $results[] = DiagnosticResult::informationalResult('PHP SAPI', php_sapi_name());
-
-        $cliPhp = new CliPhp();
-        $binary = $cliPhp->findPhpBinary();
-        if (!empty($binary)) {
-            $binary = basename($binary);
-            $rows[] = array(
-                'name'  => 'PHP Found Binary',
-                'value' => $binary,
-            );
-        }
-        $results[] = DiagnosticResult::informationalResult('PHP Timezone', date_default_timezone_get());
-        $results[] = DiagnosticResult::informationalResult('PHP Time', time());
-        $results[] = DiagnosticResult::informationalResult('PHP Datetime', Date::now()->getDatetime());
-
-        $disabled_functions = ini_get('disable_functions');
-        if (!empty($disabled_functions)) {
-            $results[] = DiagnosticResult::informationalResult('PHP Disabled functions', $disabled_functions);
-        }
-
-        foreach (['max_execution_time', 'post_max_size', 'max_input_vars', 'zlib.output_compression'] as $iniSetting) {
-            $results[] = DiagnosticResult::informationalResult('PHP INI ' . $iniSetting, @ini_get($iniSetting));
-        }
-
-        if ( function_exists( 'curl_version' ) ) {
-            $curl_version = curl_version();
-            $curl_version = $curl_version['version'] . ', ' . $curl_version['ssl_version'];
-            $results[] = DiagnosticResult::informationalResult('Curl Version', $curl_version);
-        }
-        $suhosin_installed = ( extension_loaded( 'suhosin' ) || ( defined( 'SUHOSIN_PATCH' ) && constant( 'SUHOSIN_PATCH' ) ) );
-
-        $results[] = DiagnosticResult::informationalResult('Suhosin Installed', $suhosin_installed);
-
-        if (!empty($_SERVER['HTTP_USER_AGENT'])) {
-            $results[] = DiagnosticResult::informationalResult('User Agent', $_SERVER['HTTP_USER_AGENT']);
-        }
-        $results[] = DiagnosticResult::informationalResult('Browser Language', Common::getBrowserLanguage());
-
-        if (SettingsPiwik::isMatomoInstalled()){
-            $dbConfig = Config::getInstance()->database;
-            $results[] = DiagnosticResult::informationalResult('DB Prefix', $dbConfig['tables_prefix']);
-            $results[] = DiagnosticResult::informationalResult('DB Charset', $dbConfig['charset']);
-            $results[] = DiagnosticResult::informationalResult('DB Adapter', $dbConfig['adapter']);
-            try {
-                $results[] = DiagnosticResult::informationalResult('MySQL Version', Db::get()->getServerVersion());
-            } catch (\Exception $e) {
-                $results[] = DiagnosticResult::informationalResult('MySQL Version', $e->getMessage());
-            }
-            $results[] = DiagnosticResult::informationalResult('Num Tables', $this->getNumMatomoTables());
-
-            $pluginsActivated = array();
-            $pluginsDeactivated = array();
-            $pluginsInvalid = array();
-            $plugins = Manager::getInstance()->loadAllPluginsAndGetTheirInfo();
-            foreach ($plugins as $pluginName => $plugin) {
-                $string = $pluginName;
-                if (!empty($plugin['info']['version'])
-                    && !empty($plugin['uninstallable'])
-                    && (!defined('PIWIK_TEST_MODE') || !PIWIK_TEST_MODE)) {
-                    // we only want to show versions for plugins not shipped with core
-                    // in tests we don't show version numbers to not always needing to update the screenshot
-                    $string .= ' ' . $plugin['info']['version'];
-                }
-                if (!empty($plugin['activated'])) {
-                    $pluginsActivated[] = $string;
-                } else {
-                    $pluginsDeactivated[] = $string;
-                }
-                if (!empty($plugin['invalid'])) {
-                    $pluginsInvalid[] = $string;
-                }
-            }
-
-            $results[] = DiagnosticResult::informationalResult('Plugins Activated', implode(', ', $pluginsActivated));
-            $results[] = DiagnosticResult::informationalResult('Plugins Deactivated', implode(', ', $pluginsDeactivated));
-            $results[] = DiagnosticResult::informationalResult('Plugins Invalid', implode(', ', $pluginsInvalid));
-
-            if (!empty($GLOBALS['MATOMO_PLUGIN_DIRS'])) {
-                $results[] = DiagnosticResult::informationalResult('Custom Plugins Directories', json_encode($GLOBALS['MATOMO_PLUGIN_DIRS']));
-            }
-
-            $results[] = DiagnosticResult::informationalResult('Matomo Install Version', DbHelper::getInstallVersion());
-            $results[] = DiagnosticResult::informationalResult('Had visits in last 1 day', $this->hadVisitsInLastDays(1));
-            $results[] = DiagnosticResult::informationalResult('Had visits in last 3 days', $this->hadVisitsInLastDays(3));
-            $results[] = DiagnosticResult::informationalResult('Had visits in last 5 days', $this->hadVisitsInLastDays(5));
-
-            $results[] = DiagnosticResult::informationalResult('Browser Archiving Enabled', Rules::isBrowserTriggerEnabled());
-            $results[] = DiagnosticResult::informationalResult('Browser Segment Archiving Enabled',  Rules::isBrowserArchivingAvailableForSegments());
-            $results[] = DiagnosticResult::informationalResult('Development Mode Enabled', Development::isEnabled());
-            $results[] = DiagnosticResult::informationalResult('Internet Enabled',SettingsPiwik::isInternetEnabled());
-            $results[] = DiagnosticResult::informationalResult('Multi Server Environment', SettingsPiwik::isMultiServerEnvironment());
-            $results[] = DiagnosticResult::informationalResult('Archive Time Last Started', Option::get(CronArchive::OPTION_ARCHIVING_STARTED_TS));
-            $results[] = DiagnosticResult::informationalResult('Archive Time Last Finished', Option::get(CronArchive::OPTION_ARCHIVING_FINISHED_TS));
-            $results[] = DiagnosticResult::informationalResult('Custom User Path', PIWIK_USER_PATH != PIWIK_DOCUMENT_ROOT);
-            $results[] = DiagnosticResult::informationalResult('Custom Include Path', PIWIK_INCLUDE_PATH != PIWIK_DOCUMENT_ROOT);
-        }
-        return $results;
-    }
-
-    private function getNumMatomoTables() {
-        $prefix = Common::prefixTable('');
-
-        $results = null;
-        try {
-            $results = Db::get()->fetchAll('show tables like "'.$prefix.'%"');
-        } catch (\Exception $e) {
-            return 'show tables not working';
-        }
-
-        return count($results);
-    }
-
-    private function hadVisitsInLastDays($numDays)
-    {
-        $table = Common::prefixTable('log_visit');
-        $time = Date::now()->subDay($numDays)->getDatetime();
-
-        try {
-            $row = Db::fetchOne('SELECT count(idsite) from ' . $table . ' where visit_last_action_time > ? LIMIT 1', $time );
-        } catch ( \Exception $e ) {
-            $row = null;
-        }
-
-        if (!empty($row)) {
-            return '1';
-        }
-        return '0';
+        return $this->informationalResults;
     }
 
     private function computeErrorAndWarningCount()
